[RFC]ASoC: topology: allow dapm widgets to have multiple types of kcontrols
In the current implementation, ALSA topology allows dapm
widgets to have multiple kcontrols but all of these have to be of
the same type (enum, mixer or bytes). This patch proposes to change
that and allow dapm widgets to have multiple kcontrols of different types.

One scenario where the need arises for this is the case of Equalizer.
An equalizer could potentially need a bytes control to send algorithm
data from userspace while also requiring an enum control to allow users to
switch responses dynamically at run-time.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>