ASoC: SOF: Add MFD support
Add MFD support in SOF with a single client device for audio.
The audio client is a registered as a separate platform device
under the SOF PCI device. This patch deals with partitioning
the DSP hardware initialisation parts and the audio specific
code. The hardware initialization, FW load/boot, IPC init
is done at the top-level PCI device probe. The audio client
platform device's probe deals with registering the component
driver, creating the machine driver and loading the topology.

The device hierarchy remains the same as before except for the
addition of an additional platform device for the SOF audio client.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>