ASoC: SOF: Separate SSP and HDA audio into different clients
This PR justifies the use of MFD in SOF by separating the SSP
and HDA audio in SOF into separate clients. The result of
this change is that there will 2 separate audio cards created,
one for SSP and the other for HD-audio codecs and DMIC's.

The core registers both the SSP and the HDA clients. The
sof-ssp-audio client handles the SSP dai drv's in case
there's an ACPI match found. It also handles the nocodec
mode setup. In the case of nocodec mode, the sof-ssp-audio
registers the DMIC dai drv's too.

The sof-hdmi-audio handles the HDA and DMIC dai drv
there is are HD audio codecs found on the platform.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>