ASoC: SOF: partition audio parts from core
Move all the audio-specific code in the core,
top-level PM callbacks and the core header files
to sof-audio.* in preparation for partitioning
the audio-related code for the audio client.

Also, use the component device instead of the snd_sof_dev
device wherever possible in the PCM component driver,
control IO functions and the topology parser.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>