ASoC: SOF: pm: Add flag to indicate if DSP should be put in D0i3
With the introduction of a separate client for audio, the
SOF core will no longer have access to the list of PCM's
that will be created by the client. So, instead of
interating through the list of PCM's to determine
if a D0i3-compatible stream has ignored suspend,
use the d0i3_on_suspend to determine if the DSP
should be put in D0i3. This flag will be set when
a D0i3-compatible stream receives the suspend trigger.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>