ASoC: SOF: partition audio-related parts from SOF core
Move all the audio-specific code in the core,
top-level PM callbacks and the core header files
into a separate file (sof-audio.*) in preparation for
adding an audio client device.

In the process of moving all structure definitions
for widget, routes, pcm's etc, the snd_sof_dev
member in all these structs is replaced with
the snd_soc_component member. Also, use the component
device instead of the snd_sof_dev device wherever
possible in the PCM component driver,
control IO functions and the topology parser as the
component device will be moved over to the client
device later on.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>