ASoC: SOF: topology: Add routes for PCM streams
Dynamic loading/unloading of pipelines requires that
the widgets and routes corresponding to a PCM stream be
setup only when the stream is active. The SOF driver
currently caches a list of widgets, routes
and dai configs. Identifying the widgets and the connections
for a pipeline at run-time can be quite intensive and introduce
a long latency.

This patch introduces a helper that iterates through the widgets and
routes lists and saves the information for each PCM stream at the
end of topology parsing. This information will be used to set-up/destroy
the widgets and routes for each pipeline when the PCM is opened/closed.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>